refactoring external service with super and child classes
