+++ src/main/java/codeemoji/inlay/external/BaseExternalServiceForVulnerability.java
@@ -1,0 +1,71 @@
+package codeemoji.inlay.external;
+
+import codeemoji.core.external.CEExternalService;
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.roots.libraries.Library;
+import com.intellij.openapi.roots.libraries.LibraryTable;
+import com.intellij.openapi.roots.libraries.LibraryTablesRegistrar;
+import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.psi.PsiElement;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.*;
+
+public abstract class BaseExternalServiceForVulnerability implements CEExternalService<VirtualFile, Object> {
+    protected final Map<VirtualFile, Object> persistedData = new HashMap<>();
+    protected Map<DependencyInfo, List<VulnerabilityInfo>> vulnerabilityMap = new HashMap<>();
+    protected final DependencyExtractor dependencyExtractor = new DependencyExtractor();
+    protected DependencyInfo[] dependencyInfos;
+
+    protected abstract VulnerabilityScanner getVulnerabilityScanner();
+
+    @Override
+    public void preProcess(@NotNull Project project) {
+        persistedData.put(project.getWorkspaceFile(), null);
+        dependencyInfos = getProjectLibraries(project);
+        vulnerabilityMap = getVulnerabilityScanner().scanVulnerability(dependencyInfos);
+    }
+
+    public DependencyInfo[] getProjectLibraries(Project project) {
+        LibraryTable librarytable = LibraryTablesRegistrar.getInstance().getLibraryTable(project);
+        Library[] librariesList = librarytable.getLibraries();
+        return Arrays.stream(librariesList)
+                .map(dep -> {
+                    try {
+                        return dependencyExtractor.getDependecyInfo(dep);
+                    } catch (IllegalArgumentException e) {
+                        System.err.println("Error processing dependency: " + dep.getName() + ". Error: " + e.getMessage());
+                        return null;
+                    }
+                })
+                .filter(Objects::nonNull)
+                .toArray(DependencyInfo[]::new);
+    }
+
+    @Override
+    public Map<VirtualFile, Object> getPersistedData() {
+        return persistedData;
+    }
+
+    @SuppressWarnings({"unchecked", "rawtypes"})
+    @Override
+    public void buildInfo(@NotNull Map infoResult, @Nullable PsiElement element) {
+        if (element != null && element.getProject() != null) {
+            DependencyInfo[] currentDependencies = getProjectLibraries(element.getProject());
+            boolean changes = !Arrays.equals(dependencyInfos, currentDependencies);
+            if (changes) {
+                dependencyInfos = currentDependencies;
+                vulnerabilityMap = getVulnerabilityScanner().scanVulnerability(dependencyInfos);
+            }
+        }
+        for (Map.Entry<DependencyInfo, List<VulnerabilityInfo>> entry : vulnerabilityMap.entrySet()) {
+            infoResult.put(entry.getKey(), entry.getValue());
+        }
+    }
+
+    @Nullable
+    public Object retrieveData(VirtualFile file) {
+        return persistedData.get(file);
+    }
+}
